+++ tests/integration_test.rs
@@ -18,9 +18,28 @@ fn get_compressed_file_content() {
     assert_eq!(written, 18, "Uncompressed bytes count did not match");
 }
 
+#[async_std::test]
+#[cfg(feature = "futures_support")]
+async fn get_compressed_file_content_futures() {
+    let mut source = async_std::fs::File::open("tests/fixtures/file.txt.gz")
+        .await
+        .unwrap();
+    let mut target = Vec::default();
+
+    let written = futures_support::uncompress_data(&mut source, &mut target)
+        .await
+        .expect("Failed to uncompress the file");
+    assert_eq!(
+        String::from_utf8_lossy(&target),
+        "some_file_content\n",
+        "Uncompressed file did not match",
+    );
+    assert_eq!(written, 18, "Uncompressed bytes count did not match");
+}
+
 #[tokio::test]
 #[cfg(feature = "tokio_support")]
-async fn async_get_compressed_file_content() {
+async fn get_compressed_file_content_tokio() {
     let mut source = tokio::fs::File::open("tests/fixtures/file.txt.gz")
         .await
         .unwrap();
@@ -52,9 +71,29 @@ fn get_a_file_from_tar() {
     assert_eq!(written, 14, "Uncompressed bytes count did not match");
 }
 
+#[async_std::test]
+#[cfg(feature = "futures_support")]
+async fn get_a_file_from_tar_futures() {
+    let mut source = async_std::fs::File::open("tests/fixtures/tree.tar")
+        .await
+        .unwrap();
+    let mut target = Vec::default();
+
+    let written =
+        futures_support::uncompress_archive_file(&mut source, &mut target, &"tree/branch2/leaf")
+            .await
+            .expect("Failed to get the file");
+    assert_eq!(
+        String::from_utf8_lossy(&target),
+        "Goodbye World\n",
+        "Uncompressed file did not match",
+    );
+    assert_eq!(written, 14, "Uncompressed bytes count did not match");
+}
+
 #[tokio::test]
 #[cfg(feature = "tokio_support")]
-async fn async_get_a_file_from_tar() {
+async fn get_a_file_from_tar_tokio() {
     let mut source = tokio::fs::File::open("tests/fixtures/tree.tar")
         .await
         .unwrap();
@@ -89,9 +128,29 @@ fn successfully_list_archive_files() {
     );
 }
 
+#[async_std::test]
+#[cfg(feature = "futures_support")]
+async fn successfully_list_archive_files_futures() {
+    let source = async_std::fs::File::open("tests/fixtures/tree.tar")
+        .await
+        .unwrap();
+
+    assert_eq!(
+        futures_support::list_archive_files(source).await.unwrap(),
+        vec![
+            "tree/".to_string(),
+            "tree/branch1/".to_string(),
+            "tree/branch1/leaf".to_string(),
+            "tree/branch2/".to_string(),
+            "tree/branch2/leaf".to_string(),
+        ],
+        "file list inside the archive did not match"
+    );
+}
+
 #[tokio::test]
 #[cfg(feature = "tokio_support")]
-async fn async_successfully_list_archive_files() {
+async fn successfully_list_archive_files_tokio() {
     let source = tokio::fs::File::open("tests/fixtures/tree.tar")
         .await
         .unwrap();
@@ -251,9 +310,36 @@ fn uncompress_same_file_not_preserve_owner() {
     .expect("Failed to uncompress the file");
 }
 
+#[async_std::test]
+#[cfg(feature = "futures_support")]
+async fn uncompress_same_file_not_preserve_owner_futures() {
+    futures_support::uncompress_archive(
+        &mut async_std::fs::File::open("tests/fixtures/tree.tar")
+            .await
+            .unwrap(),
+        tempfile::TempDir::new()
+            .expect("Failed to create the tmp directory")
+            .path(),
+        Ownership::Ignore,
+    )
+    .await
+    .expect("Failed to uncompress the file");
+    futures_support::uncompress_archive(
+        &mut async_std::fs::File::open("tests/fixtures/tree.tar")
+            .await
+            .unwrap(),
+        tempfile::TempDir::new()
+            .expect("Failed to create the tmp directory")
+            .path(),
+        Ownership::Ignore,
+    )
+    .await
+    .expect("Failed to uncompress the file");
+}
+
 #[tokio::test]
 #[cfg(feature = "tokio_support")]
-async fn async_uncompress_same_file_not_preserve_owner() {
+async fn uncompress_same_file_not_preserve_owner_tokio() {
     tokio_support::uncompress_archive(
         &mut tokio::fs::File::open("tests/fixtures/tree.tar")
             .await
@@ -289,9 +375,24 @@ fn uncompress_truncated_archive() {
     ));
 }
 
+#[async_std::test]
+#[cfg(feature = "futures_support")]
+async fn uncompress_truncated_archive_futures() {
+    assert!(matches!(
+        futures_support::uncompress_data(
+            async_std::fs::File::open("tests/fixtures/truncated.log.gz")
+                .await
+                .unwrap(),
+            Vec::new()
+        )
+        .await,
+        Err(Error::Unknown)
+    ));
+}
+
 #[tokio::test]
 #[cfg(feature = "tokio_support")]
-async fn async_uncompress_truncated_archive() {
+async fn uncompress_truncated_archive_tokio() {
     assert!(matches!(
         tokio_support::uncompress_data(
             tokio::fs::File::open("tests/fixtures/truncated.log.gz")