+++ CHANGES.md
@@ -4,8 +4,8 @@
 
 ### Added
 
-* Optional tokio support
-* Uncompress service example and its async counterpart
+* Optional async support
+* Uncompress service example and its async-std and Tokio counterparts
 
 ### Changed